@@ -1,7 +1,8 @@
 cmake_minimum_required(VERSION 3.1)
 project(Eng)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if(MSVC)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /MP")
@@ -353,7 +354,7 @@ else(WIN32)
         target_link_directories(Eng PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/third-party/vpx/linux/x64)
     endif(APPLE)
 endif(WIN32)
-target_link_libraries(Eng Net Phy Ren Snd Sys astc optick ${GLSLANG_LIBS} ${ITT_LIB} ${VPX_LIB})
+target_link_libraries(Eng Net Phy Ren Snd Sys astc optick glslx_lib ${GLSLANG_LIBS} ${ITT_LIB} ${VPX_LIB})
 
 add_subdirectory(tests)
 add_subdirectory(first-party/Net)
@@ -361,7 +362,9 @@ add_subdirectory(first-party/Phy)
 add_subdirectory(first-party/Ren)
 add_subdirectory(first-party/Sys)
 add_subdirectory(first-party/Snd)
+add_subdirectory(first-party/glslx)
 add_subdirectory(third-party/ltc_fit)
 add_subdirectory(third-party/astc)
 add_subdirectory(third-party/optick)
 target_include_directories(optick PUBLIC first-party/Ren/third-party)
+set_target_properties(glslx PROPERTIES EXCLUDE_FROM_ALL 1 EXCLUDE_FROM_DEFAULT_BUILD 1)
