@@ -3,7 +3,7 @@ project(Sys)
 
 set(CMAKE_CONFIGURATION_TYPES "Debug;Release;RelWithDebInfo;Asan" CACHE STRING "" FORCE)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 find_library(CORE_GRAPHICS CoreGraphics)
@@ -47,7 +47,6 @@ set(SOURCE_FILES AssetFile.h
                  Json.cpp
                  MemBuf.h
                  MonoAlloc.h
-                 Optional.h
                  Pack.h
                  Pack.cpp
                  PoolAlloc.h
@@ -59,7 +58,6 @@ set(SOURCE_FILES AssetFile.h
                  ThreadWorker.h
                  Time_.h
                  Time.cpp
-                 Variant.h
                  WindowRect.h
                  WindowRect.cpp)
 
