@@ -66,7 +66,7 @@ inline bool ThreadWorker::Stop() {
 template <class F, class... Args>
 auto ThreadWorker::AddTask(F &&f, Args &&... args)
     -> std::future<typename std::result_of<F(Args...)>::type> {
-    using return_type = typename std::result_of<F(Args...)>::type;
+    using return_type = typename std::invoke_result_t<F>;
 
     auto task = std::make_shared<std::packaged_task<return_type()>>(
         std::bind(std::forward<F>(f), std::forward<Args>(args)...));
